Modify repo structure, recreate block design and rewrite game logic
1. The repository structure
Repository folders are all lower-case and path to the SDK sources
has been simplified.

2. Block design recreation
HDL code is responsible just for sending data to a VGA monitor
and not for doing any logic operations connected with the game.
Moreover, each game element is going to have its own texture stored
in BRAM, which basically means no difference in code for drawing,
e.g., player and bomb.

Drawing means sending pixel color value to the rgb output.

Thus, code for drawing every element is the same so that all of
the elements can be handled in one Verilog module called
vga_drawer/board_draw.

Instead of using multiple AXI blocks for various elements, there's
need for only one AXI block capable of storing all of the game
elements (axi_board).
Memory write pattern:
element_type(4-bits) | element_position(8-bits)
The C++ code takes responsibility for correct management of
this memory.

A Verilog module, called vga_drawer/rom_rgb_mux, is
to choose the right rgb value to send to the output (ordinary MUX).

We also have a new AXI Timer IP for time management in the game.

Textures stored in BRAMs are plain colors. Overall BRAM usage is
89% (44.5/50) - we are very limited in this aspect.

3. Game logic
For now, we have one bomb per player to make things easier.

Every game element is drawn by BoardMemory object. Before drawing we
first need to prepare data for drawing. E.g., if a player stands on
a bomb (because he's just activated it), we want to draw this player
instead of the bomb. Elements to draw are set right before
actual drawing. If we just drew all of the elements without former
preparation, we would get colors changing fast on the screen.
This is the correct management of the memory mentioned in paragraph 2.

Player movement speed is set to 1 move per 0.3 sec.
Bomb active time is set to 4 seconds, which means a bomb is visible
on the screen for 4 seconds.

Next step is to create Explosion class and test collisions with blocks
and players. We can also create our own textures for elements.